max31855.c: Channel 0 will be the data channel, and 1 is the error register code.
myAnalogRead:
Return the analog value of the given pin
Note: The chip really only has one read "channel", but we're faking it
here so we can read the error registers. Channel 0 will be the data
channel, and 1 is the error register code. Note: Temperature returned is temp in C  4, so divide result by 4
Read temp in C Bottom 13 bits Negative Return error bits
max31855Setup:
Create a new wiringPi device node for an max31855 on the Pi's
SPI interface.