pcf8591.c: pcf8591.c: Extend wiringPi with the PCF8591 I2C GPIO Analog expander chip The chip has 1 8-bit DAC and 4 x 8-bit ADCs Copyright (c) 2013 Gordon Henderson This file is part of wiringPi: https:projects.drogon.netraspberry-piwiringpi wiringPi is free software: you can redistribute it andor modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
pcf8591.c:
Extend wiringPi with the PCF8591 I2C GPIO Analog expander chip
The chip has 1 8-bit DAC and 4 x 8-bit ADCs
Copyright (c) 2013 Gordon Henderson
This file is part of wiringPi:
https:projects.drogon.netraspberry-piwiringpi
wiringPi is free software: you can redistribute it andor modify
it under the terms of the GNU Lesser General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version. wiringPi is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. We know it
has 4 pins, (4 analog inputs and 1 analog output which we'll shadow
input 0) so all we need to know here is the I2C address and the
user-defined pin base.