softPwm.c: Another way to increase the frequency is to reduce the range - however	that reduces the overall output accuracy... softPwmThread: Thread to do the actual PWM output softPwmWrite: Write a PWM value to the given pin softPwmCreate: Create a new softPWM thread.
It's possible to get a higher frequency by lowering the pulse time,	however CPU uage will skyrocket as wiringPi uses a hard-loop to time	periods under 100µS - this is because the Linux timer calls are just	accurate at all, and have an overhead. Another way to increase the frequency is to reduce the range - however	that reduces the overall output accuracy...
softPwmThread:
Thread to do the actual PWM output
softPwmWrite:
Write a PWM value to the given pin
softPwmCreate:
Create a new softPWM thread. Already running on this pin
softPwmStop:
Stop an existing softPWM thread