softServo.c: In this code, we don't need to be too fussy about the gap as we're not doing	the multipexing, but it does need to be at least 10mS, and preferably 16	from what I've been able to determine.
softServo.c:
Provide N channels of software driven PWM suitable for RC
servo motors. So... To drive an RC Servo motor, you need to send it a modified PWM	signal - it needs anything from 1ms to 2ms - with 1ms meaning	to move the server fully left, and 2ms meaning to move it fully	right. In this code, we don't need to be too fussy about the gap as we're not doing	the multipexing, but it does need to be at least 10mS, and preferably 16	from what I've been able to determine.