sr595.c: Note that the code can cope with a number of 595's daisy-chained together - up to 4 for now as we're storing the output "register" in a single unsigned int.
Note that the code can cope with a number of 595's
daisy-chained together - up to 4 for now as we're storing
the output "register" in a single unsigned int. number of clock pulses A low -> high latch transition copies the latch to the output pins
sr595Setup:
Create a new instance of a 74x595 shift register GPIO expander. Output register Initialise the underlying hardware