wiringPiI2C.c: This may ultimately reduce the flexibility of this code, but it won't be hard to maintain it and keep it current, should things change.
Notes:
The Linux I2C code is actually the same (almost) as the SMBus code. I'm directly including the SMBus functions here as some Linux distros
lack the correct header files, and also some header files are GPLv2
rather than the LGPL that wiringPi is released under - presumably because
originally no-one expected I2CSMBus to be used outside the kernel -
however enter the Raspberry Pi with people now taking directly to I2C
devices without going via the kernel... This may ultimately reduce the flexibility of this code, but it won't be
hard to maintain it and keep it current, should things change.