wiringPiSPI.c: The SPI bus parameters	Variables as they need to be passed as pointers later on wiringPiSPIGetFd: Return the file-descriptor for the given channel wiringPiSPIDataRW: Write and Read a block of data over the SPI bus.
The SPI bus parameters	Variables as they need to be passed as pointers later on
wiringPiSPIGetFd:
Return the file-descriptor for the given channel
wiringPiSPIDataRW:
Write and Read a block of data over the SPI bus. Note the data ia being read into the transmit buffer, so will
overwrite it! I've no idea, but for now I'm blindly	copying example code I've seen online...