wiringSerial.c: serialOpen: Open and initialise the serial port, setting all the right port parameters - or as many as are required - hopefully!
wiringPi is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. serialOpen:
Open and initialise the serial port, setting all the right
port parameters - or as many as are required - hopefully! Get and modify current options: Ten seconds (100 deciseconds) 10mS
serialFlush:
Flush the serial buffers (both tx & rx)
serialClose:
Release the serial port
serialPutchar:
Send a single character to the serial port
serialPuts:
Send a string to the serial port
serialPrintf:
Printf over Serial
serialDataAvail:
Return the number of bytes of data avalable to be read in the serial port
serialGetchar:
Get a single character from the serial device.